+++ next-app/src/components/common/Scripts/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Scripts'
